auth

+++ subsystems/db/model_user.py
@@ -20,7 +20,7 @@ EXP_MAP = [68, 295, 805, 1716, 3154, 5249, 8136, 11955, 16851, 22973, 30475,
 class User(models.Model):
     # ========== user auth information ==========
     id = models.AutoField(primary_key=True)
-    id_vk = models.IntegerField(null=True)
+    id_vk = models.BigIntegerField(null=True)
     password = models.CharField(max_length=128, null=True)
     is_superuser = models.BooleanField(default=False)
     email = models.EmailField(max_length=30, unique=True, null=True)
